using local flyway in devcontainers

+++ flyway-clean.sh
@@ -1,14 +1,20 @@
 #! /usr/bin/bash
 
-
 set -e # Exit on error
+set -x # Print commands
 
 DIR=$(dirname "$(readlink -f "$0")")
 pushd $DIR
 
-docker run --rm -it --network=host \
-    -v $(pwd)/flyway-migrations:/flyway/flyway-migrations \
-    -v $(pwd)/flyway.conf:/flyway/conf/flyway.conf \
-    flyway/flyway:9.20 clean
+if command -v flyway >/dev/null 2>&1; then
+	echo "flyway command found, using it"
+	flyway clean -configFiles=flyway.conf
+else
+    echo "flyway command not found, using docker"
+    docker run --rm -it --network=host \
+        -v $(pwd)/flyway-migrations:/flyway/flyway-migrations \
+        -v $(pwd)/flyway.conf:/flyway/conf/flyway.conf \
+        flyway/flyway:9.20 clean
+else
 
 popd
