+++ flyway-migrate.sh
@@ -1,13 +1,21 @@
 #! /usr/bin/bash
 
 set -e # Exit on error
+set -x # Print commands
 
 DIR=$(dirname "$(readlink -f "$0")")
 pushd $DIR
 
-docker run --rm -it --network=host \
-	-v $(pwd)/flyway-migrations:/flyway/flyway-migrations \
-	-v $(pwd)/flyway.conf:/flyway/conf/flyway.conf \
-	flyway/flyway:9.21 migrate
+if command -v flyway >/dev/null 2>&1; then
+	echo "flyway command found, using it"
+	flyway migrate -configFiles=flyway.conf
+else
+	echo "flyway command not found, using docker"
+	docker run --rm -it --network=host \
+		-v $(pwd)/flyway-migrations:/flyway/flyway-migrations \
+		-v $(pwd)/flyway.conf:/flyway/conf/flyway.conf \
+		flyway/flyway:9.21 migrate
+fi
+
 
 popd
